% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLandmarks.R
\name{findLandmarks}
\alias{findLandmarks}
\title{findLandmarks - A function for finding landmarks within
 a resulting data frame from xcms}
\usage{
findLandmarks(
  dat,
  mzrtdf,
  mzdiff,
  rtdiff,
  minIntensity,
  minrttocheck,
  allowedmissingness
)
}
\arguments{
\item{dat}{Data frame with features in columns and samples in rows}

\item{mzrtdf}{Data frame with mz values of all features in column 1
and retention time values of all features in column 2 named mzvec and rtvec}

\item{mzdiff}{How large whitespace in mz a feature needs, typically 0.1}

\item{rtdiff}{How large whitespace in rt a feature needs,
can vary quite a lot, but ~20 is sufficient}

\item{minIntensity}{Value indicating what min intensity is needed
for a feature to qualify as an LM}

\item{minrttocheck}{Retention time value where noisy features
(that usually comes in the beginning of a run) can be ignored}

\item{allowedmissingness}{Double between 0 - 1 specifying how 
much missingness allowed for LaMas within all samples}
}
\value{
Returning potential Landmarks
}
\description{
findLandmarks - A function for finding landmarks within
 a resulting data frame from xcms
}
