% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmarks_final}
\alias{landmarks_final}
\title{Landmarks Final Dataset}
\format{
A data frame containing mz and rt for all landmarks where each row is
a landmark
\describe{
  \item{MZ}{Mass-to-charge ratio of landmark features}
  \item{RT}{Retention time of landmark features}
  ...
}
}
\usage{
data(landmarks_final)
}
\description{
Landmarks found through use of the 'optimizeLM()' function performed on a
peak table derived from xcms preprocessing up until, but not including,
gap filling
}
\keyword{datasets}
