% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeFindLM.R
\name{optimizeFindLM}
\alias{optimizeFindLM}
\title{optimizeFindLM - A function for optimization of finding new LaMas}
\usage{
optimizeFindLM(
  PT_filepath = NULL,
  PT = NULL,
  split = "@",
  minrttocheck = 40,
  Prefilterintensity,
  minLM = 100,
  maxLM = 200
)
}
\arguments{
\item{PT_filepath}{File path to a .rds file which contains a peak table
generated from an XCMS object which has been processed until,
but not including, the gap-filling step}

\item{PT}{A peak table generated from an XCMS object which has been
processed until, but not including, the gap-filling step}

\item{split}{The character used in the XCMS object for splitting up mz & RT}

\item{minrttocheck}{Dead volume cut-off in beginning of chromatogram to 
disregard when looking for LaMas}

\item{Prefilterintensity}{Prefilter intensity settings used when peak 
picking with XCMS}

\item{minLM}{The minimum number of landmarks which are required to be
returned by the optimization in order to count as a result}

\item{maxLM}{The maxmimum number of landmarks allowed to be included in
the optimization}
}
\value{
Returning a list containing 10 sets of optimized LaMas, the result
parameters and a df of the result parameters
}
\description{
optimizeFindLM - A function for optimization of finding new LaMas
}
\examples{
data("peak_table")
library(rtRecalibrate)

landmark_suggestions <- optimizeFindLM(PT=peak_table,
                                       split='@',
                                       Prefilterintensity=7000,
                                      minLM = 40)

landmarks_final <- landmark_suggestions[[1]][[2]]

print(landmarks_final)
}
