% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peak_table}
\alias{peak_table}
\title{Peak Table Dataset}
\format{
A data frame peak table with sample in rows and features in columns
where mz and rt
}
\usage{
data(peak_table)
}
\description{
Peak table generated in xcms using IPO and IPO2 to find optimized parameters
for xmcs functions: findChromPeaks, adjustRtime, groupChromPeaks and finally
featureValues and featureDefinitions to build the peak table
}
\keyword{datasets}
