% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTAdjust.R
\name{plotRTAdjust}
\alias{plotRTAdjust}
\title{Plot retention time (RT) adjustment}
\usage{
plotRTAdjust(raw, adj, col, legend = TRUE, cex.legend = 0.5)
}
\arguments{
\item{raw}{Original data (in MsExperiment or MSnbase format)}

\item{adj}{RT-adjusted data (in same format as raw)}

\item{col}{Optional argument to specify plot color
(should correspond to the order of files in raw and adj)}

\item{legend}{Optional argument to plot (TRUE; default)
or suppress (FALSE) the legend}

\item{cex.legend}{Optional argument to change font size in legend}
}
\value{
Returns a plot of RT adjustment vs the original recorded RTs
}
\description{
Plot retention time (RT) adjustment
}
