% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_rt.R
\name{recalibrateRt}
\alias{recalibrateRt}
\alias{recalibrateRt,MsExperiment-method}
\alias{recalibrateRt,OnDiskMSnExp-method}
\title{Recalibrate retention times based on landmark compounds}
\usage{
recalibrateRt(object, ...)

\S4method{recalibrateRt}{MsExperiment}(
  object,
  lamas,
  method = c("GAM", "loess"),
  roiExpandRt = 15,
  roiPpm = 40,
  toleranceRt = 15,
  ppm = 5,
  tolerance = 0,
  peakwidth = c(7, 50),
  bs = "tp",
  span = 0.5,
  ssqRatio = 3,
  zeroWeight = 10,
  BPPARAM = bpparam()
)

\S4method{recalibrateRt}{OnDiskMSnExp}(
  object,
  lamas,
  method = c("GAM", "loess"),
  roiExpandRt = 15,
  roiPpm = 40,
  toleranceRt = 15,
  ppm = 5,
  tolerance = 0,
  peakwidth = c(7, 50),
  bs = "tp",
  span = 0.5,
  ssqRatio = 3,
  zeroWeight = 10,
  BPPARAM = bpparam()
)
}
\arguments{
\item{object}{Either a `MsExperiment` or `OnDiskMSnExp` object with the MS
data.}

\item{...}{ignored.}

\item{lamas}{`data.frame` with m/z values (first column) and retention
times (second columns) of the landmarks.}

\item{method}{`character(1)` with the type of warping. Either
`method = "GAM"` (default) or `method = "loess"`.}

\item{roiExpandRt}{`numeric(1)` defining the retention time width left and
right of the landmark's expected retention time (second column in
`lamas`) where peak detection should be performed.}

\item{roiPpm}{`numeric(1)` defining the m/z width around the landmark's
expected m/z (first column in `lamas`) where peak detection should be
performed.}

\item{toleranceRt}{`numeric(1)` with the maximal allowed (absolute)
difference in retention times between lamas and chromatographic peaks.
Used for the mapping of identified chromatographic peaks and lamas.}

\item{ppm}{`numeric(1)` defining the m/z-relative maximal allowed
difference in m/z between lamas and chromatographic peaks. Used for
the mapping of identified chromatographic peaks and lamas.}

\item{tolerance}{`numeric(1)` defining the absolute acceptable difference
in m/z between lamas and chromatographic peaks. Used for the mapping
of identified chromatographic peaks and lamas.}

\item{peakwidth}{`numeric(2)` with the lower and upper limit (in retention
time) for the *centWave*-based chromatographic peak detection.}

\item{bs}{`character(1)` defining the GAM moothing method. (defaults to
thin plate; NB: B- and P-splines have been shown to produce artefacts).}

\item{span}{`numeric(1)` loess smoothing parameter.}

\item{ssqRatio}{`numeric(1)` defining the settings for removing outliers
in the warping.}

\item{zeroWeight}{`numeric(1)`: additional weight to the zero point {0,0},
reduces warping problems in the early RT range.}

\item{BPPARAM}{Parallel processing setup. See [bpparam()] for more
information.}
}
\value{
Input object with recalibrated retention times.
}
\description{
Recalibrate retention times of provided MS data based on landmark features
(lamas). For input data `object` being an `MsExperiment` (from the
*MsExperiment* package) or `OnDiskMSnExp` (from the *MSnbase* package)
chromatographic peak detection will be performed specifically looking for
the landmark features and the retention times of the full data set will
be recalibrated by aligning the observed retention times of the landmark
features to the provided *expected* ones. Chromatographic peak detection
will be performed using the *centWave* method from the *xcms* package.

Chromatographic peaks are matched to lamas based on their m/z and retention
times and parameters `toleranceRt`, `ppm` and `tolerance` defining
acceptable deviations. Only lamas/chromatographic peaks with a 1:1 mapping
are used to estimate retention time shifts.
}
\author{
Carl Brunius, Johannes Rainer.
}
