% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtRecal.R
\name{rtRecal}
\alias{rtRecal}
\title{Recalibrate retention time axis of mzML file to landmark reference
 - One file only}
\usage{
rtRecal(
  file,
  lamas,
  method = c("GAM", "loess"),
  dRT_roi = 15,
  ppm_roi = 40,
  dRT_match = 15,
  ppm_match = 5,
  peakwidth = c(7, 50),
  expandRt = 0,
  bs = "tp",
  span = 0.5,
  ssqRatio = 3,
  zeroWeight = 10,
  jpg = TRUE,
  plot = TRUE,
  save = TRUE
)
}
\arguments{
\item{file}{Path to a mzML file}

\item{lamas}{Data frame with 2 columns describing the landmarks: \"MZ\" in 
the first and \"RT\" in the second}

\item{method}{Type of warping performed: GAM or loess}

\item{dRT_roi}{Determines RT width of ROIs for detecting landmarks}

\item{ppm_roi}{Determines m/z width (in ppm) of ROIs for detecting landmarks}

\item{dRT_match}{Tolerance (dRT) for matching possible ROI peaks to landmarks}

\item{ppm_match}{Tolerance (dRT) for matching possible ROI peaks to landmarks}

\item{peakwidth}{CentWave parameter to XCMS for detecting peaks in ROIs}

\item{expandRt}{Defines the retention time range of the
ROIs in lama_peak_detection(). The retention time range will be the 
expected retention time of a lama +/- `expandRt`.}

\item{bs}{GAM smoothing method (defaults to thin plate; NB: B- and P-splines 
have been shown to produce artefacts)}

\item{span}{Loess smoothing parameter}

\item{ssqRatio}{Parameter for removing outliers in the warping}

\item{zeroWeight}{Adding additional weight to the zero point \{0,0\}, 
reduces "the"weird" warping in the early RT range}

\item{jpg}{Whether to store diagnostic plot in jpg format in 
"rtRecal_log/*.jpg" (TRUE/FALSE)}

\item{plot}{Whether to produce plots (TRUE/FALSE; if jpg is TRUE, plots go 
to file instead of plot window)}

\item{save}{Whether to save recalibrated file in "rtRecal/*.mzML"}
}
\value{
Depends of save, plot and jpg
}
\description{
Please see rtRecalibrate for standard interface for
 multiple files and parallelization
}
