% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtRecalibratemzML.R
\name{rtRecalibratemzML}
\alias{rtRecalibratemzML}
\title{Recalibrate retention time axis of mzML file}
\usage{
rtRecalibratemzML(
  files,
  lamas,
  method = c("GAM", "loess"),
  dRT_roi = 15,
  ppm_roi = 40,
  dRT_match = 15,
  ppm_match = 5,
  peakwidth = c(7, 50),
  bs = "tp",
  span = 0.5,
  ssqRatio = 3,
  zeroWeight = 10,
  jpg = TRUE,
  plot = TRUE,
  save = TRUE,
  ...
)
}
\arguments{
\item{files}{Path to mzML file(s)}

\item{lamas}{Data frame with 2 columns describing the landmarks:
m/z in the first and RT in the second}

\item{method}{Type of warping performed: GAM or loess}

\item{dRT_roi}{Determines RT width of ROIs for detecting landmarks}

\item{ppm_roi}{Determines m/z width (in ppm) of ROIs for detecting landmarks}

\item{dRT_match}{Tolerance (dRT) for matching possible ROI peaks to landmarks}

\item{ppm_match}{Tolerance (dPPM) for matching
possible ROI peaks to landmarks}

\item{peakwidth}{CentWave parameter to XCMS for detecting peaks in ROIs}

\item{bs}{GAM smoothing method (defaults to thin plate;
NB: B- and P-splines have been shown to produce artefacts)}

\item{span}{Loess smoothing parameter}

\item{ssqRatio}{Parameter for removing outliers in the warping}

\item{zeroWeight}{Adding additional weight to the zero point \{0,0\},
reduces "the"weird" warping in the early RT range}

\item{jpg}{Whether to store diagnostic plot in jpg format
in "rtRecal_log/*.jpg" (TRUE/FALSE)}

\item{plot}{Whether to produce plots
(TRUE/FALSE; if jpg is TRUE, plots go to file instead of plot window)}

\item{save}{Whether to save recalibrated file in "rtRecal/*.mzML"}

\item{...}{Additional arguments}
}
\value{
Depends of save, plot and jpg
}
\description{
Recalibrate retention time axis of mzML file
}
\examples{
data("landmarks_final")
\donttest{
  if (requireNamespace("MsBackendMetaboLights", quietly = TRUE)) {
    MsBackendMetaboLights::mtbls_sync_data_files(
      mtblsId = "MTBLS8735",
      assayName = "a_MTBLS8735_LC-MS_positive_hilic_metabolite_profiling.txt"
    )
    
    mzMLFiles <- MsBackendMetaboLights::mtbls_cached_data_files()$rpath
    
    res_mse <- rtRecalibrate::rtRecalibratemzML(files=mzMLFiles,
                                   lamas=landmarks_final,
                                   method="GAM")
  }
}

                      
                                        
}
